Add files via upload
Programs used to model how to fine phase the SC linac.  Either a set of points around the zero crossings or a full 360 deg set of data.  Random BPM noise is added to each measurement prior to fitting and the BPM resolution, dispersion and phasing beam energy are hand selected.